Added 2023 wk14
Lots of work with date functions.

Used a macro out of dbt utils to union a list of tables specified in some Jinja.

Also used Jinja to set an initial query parameter to read in before the script to set weeks to start on monday rather than the default Sunday.